\name{MicroBVS-package}
\alias{MicroBVS-package}
\alias{MicroBVS}
\docType{package}
\title{
\packageTitle{MicroBVS}
}
\description{
\packageDescription{MicroBVS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MicroBVS}
\packageIndices{MicroBVS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MicroBVS}

Maintainer: \packageMaintainer{MicroBVS}
}
\references{~~ Literature or other references for background information ~~}
\keyword{ package }
\examples{~~ simple examples of the most important functions ~~}
