\name{MicroBVS-package}
\alias{MicroBVS-package}
\alias{MicroBVS}
\docType{package}
\title{
\packageTitle{MicroBVS}
}
\description{
\packageDescription{MicroBVS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MicroBVS}
\packageIndices{MicroBVS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MicroBVS}

Maintainer: \packageMaintainer{MicroBVS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~}
